refactor: auth guard feat: passport

+++ backend/src/users/users.controller.ts
@@ -6,10 +6,14 @@ import {
   Param,
   Delete,
   ParseIntPipe,
+  NotFoundException,
+  Req,
+  ForbiddenException,
 } from "@nestjs/common";
 import { UsersService } from "./users.service";
 import { Public } from "src/public";
 import { UpdateUserDto } from "./users.types";
+import { type AuthenticatedRequest } from "src/authentication/authentication.types";
 
 @Controller("users")
 export class UsersController {
@@ -33,20 +37,59 @@ export class UsersController {
   }
 
   @Get(":id")
-  findUserById(@Param("id", ParseIntPipe) id: number) {
-    return this.usersService.findById(id);
+  async findUserById(
+    @Param("id", ParseIntPipe) id: number,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    console.log("\n\n", req);
+    if (req.user.id !== id) {
+      throw new ForbiddenException();
+    }
+
+    const user = await this.usersService.findById(id);
+
+    if (!user) {
+      throw new NotFoundException(`User with id ${id} not found`);
+    }
+
+    return user;
   }
 
   @Patch(":id")
   updateUser(
     @Param("id", ParseIntPipe) id: number,
+    @Req() req: AuthenticatedRequest,
     @Body() updateUserDto: UpdateUserDto,
   ) {
+    if (req.user.id !== id) {
+      throw new ForbiddenException();
+    }
+
     return this.usersService.update(id, updateUserDto);
   }
 
+  @Patch(":id/password")
+  updatePassword(
+    @Param("id", ParseIntPipe) id: number,
+    @Req() req: AuthenticatedRequest,
+    @Body() password: string,
+  ) {
+    if (req.user.id !== id) {
+      throw new ForbiddenException();
+    }
+
+    return this.usersService.updatePassword(id, password);
+  }
+
   @Delete(":id")
-  deleteUser(@Param("id", ParseIntPipe) id: number) {
+  deleteUser(
+    @Param("id", ParseIntPipe) id: number,
+    @Req() req: AuthenticatedRequest,
+  ) {
+    if (req.user.id !== id) {
+      throw new ForbiddenException();
+    }
+
     return this.usersService.remove(id);
   }
 }
